backend/pyproject.toml
[project]
name = "alter-ego-backend"
version = "0.1.0"
description = "Alter Ego - AI Influencer Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.5",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.3",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
